[MINDEXER-137] Align example packages to core
The examples packages used o.a.m.indexER vs o.a.m.index of core.

---

https://issues.apache.org/jira/browse/MINDEXER-137
