+++ indexer-core/src/test/mindexer-35/org/apache/maven/indexer/test/sample-war/maven-metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <metadata>
-  <groupId>org.apache.maven.indexer.test</groupId>
+  <groupId>org.apache.maven.index.test</groupId>
   <artifactId>sample-war</artifactId>
   <versioning>
     <versions>
