+++ indexer-examples/indexer-examples-basic/pom.xml
@@ -49,11 +49,16 @@ under the License.
     UNWORKABLE! -->
 
   <dependencies>
-      <!-- For search app in general -->
+      <!-- Logging -->
       <dependency>
         <groupId>org.slf4j</groupId>
         <artifactId>slf4j-api</artifactId>
         <scope>compile</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-simple</artifactId>
+        <scope>test</scope>
       </dependency>
 
       <dependency>
@@ -84,11 +89,6 @@ under the License.
       <dependency>
         <groupId>junit</groupId>
         <artifactId>junit</artifactId>
-        <scope>test</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.slf4j</groupId>
-        <artifactId>slf4j-simple</artifactId>
         <scope>test</scope>
       </dependency>
   </dependencies>
