+++ indexer-examples/indexer-examples-spring/src/main/resources/META-INF/spring/maven-indexer-context.xml
@@ -26,7 +26,7 @@
                            http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.0.xsd
                            http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.0.xsd">
 
-    <context:component-scan base-package="org.apache.maven.indexer.examples"/>
+    <context:component-scan base-package="org.apache.maven.index.examples"/>
 
     <!-- This context contains the maven-indexer beans and the configuration of the indexing. -->
